fix for shared DetectEncryptionAlgorithm for kubeconfig
